+++ lib/PhpParser/Node/Stmt/Namespace_.php
@@ -6,6 +6,10 @@ use PhpParser\Node;
 
 class Namespace_ extends Node\Stmt
 {
+    /* For use in the "kind" attribute */
+    const KIND_SEMICOLON = 1;
+    const KIND_BRACED = 2;
+
     /** @var null|Node\Name Name */
     public $name;
     /** @var Node[] Statements */